requirements.txt
python-dotenv>=1.0.0
jinja2>=3.1.2
flask>=3.0.0
flask-cors>=4.0.0
werkzeug>=3.0.0
